Fix 缩略图生成函数中加载 Clip 问题；Update 管理界面 UI
- VideoResourceManager: UI 增加刷新按钮等
- SelectPanel: 更新添加图片时等待 Sprite 加载完成

+++ Assets/Scripts/UI/VideoResourceManager.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
+using UnityEngine.UI;
 using UnityEngine.Video;
 
 public enum VideoType
@@ -14,11 +15,18 @@ public enum VideoType
 public class VideoResourceManager : MonoBehaviour
 {
     public VideoManager video;
+    public Text SelectTagText;
     public List<SelectPanel> Panels;
+    public List<Image> ButtonImages;
+    public Color SelectBtnColor;
+    public Color NotSelectBtnColor;
 
     private int curPanelIndex = 0;
     private Dictionary<string, VideoType> Prefix2Type = new Dictionary<string, VideoType>() {
         {"BG", VideoType.Background}, {"EF", VideoType.Effect}, {"PRO", VideoType.Product}
+    };
+    private Dictionary<VideoType, string> Type2Tag = new Dictionary<VideoType, string>() {
+        {VideoType.Background, "背景"}, {VideoType.Effect, "特效"}, { VideoType.Product, "成品"}
     };
 
 
@@ -29,6 +37,18 @@ public class VideoResourceManager : MonoBehaviour
         for(int i = 1; i < Panels.Count; i++) {
             Panels[i].gameObject.SetActive(false);
         }
+
+        // 按钮效果
+        if (ButtonImages.Count > 0) {
+            ButtonImages[0].color = SelectBtnColor;
+        }
+        for (int i = 1; i < ButtonImages.Count; i++) {
+            ButtonImages[i].color = NotSelectBtnColor;
+        }
+        // Tag Text
+        if (SelectTagText) {
+            SelectTagText.text = Type2Tag[VideoType.Background];
+        }
     }
 
     // 按钮事件 切换 Panel
@@ -36,6 +56,16 @@ public class VideoResourceManager : MonoBehaviour
         if (index == curPanelIndex) return;
         Panels[curPanelIndex].gameObject.SetActive(false);
         Panels[index].gameObject.SetActive(true);
+
+        // 按钮效果
+        if (index < ButtonImages.Count) {
+            ButtonImages[curPanelIndex].color = NotSelectBtnColor;
+            ButtonImages[index].color = SelectBtnColor;
+        }
+        // Tag 标签 Text
+        if (SelectTagText) {
+            SelectTagText.text = Type2Tag[(VideoType)index];
+        }
         curPanelIndex = index;
     }
 
@@ -46,8 +76,13 @@ public class VideoResourceManager : MonoBehaviour
     }
 
     // 按钮事件 删除视频
+    private VideoPlayer videoPlayer;
     public void DeleteVideo() {
-        string name = video.GetComponent<VideoPlayer>().clip.name;
+        if (videoPlayer == null)
+            videoPlayer = video.GetComponent<VideoPlayer>();
+        if (videoPlayer == null || videoPlayer.clip == null)
+            return;
+        string name = videoPlayer.clip.name;
         string prefix = name.Split('_')[0];
         string path = Prefix2Type[prefix].ToString("g");
         VideoResourceAPI.Delete(path, name);
@@ -55,5 +90,10 @@ public class VideoResourceManager : MonoBehaviour
         video.Stop();
         Panels[(int)Prefix2Type[prefix]].UpdateAll();
     }
+    
 
+    // 按钮事件 刷新当前Panel
+    public void RefreshCurrentPanel() {
+        Panels[curPanelIndex].UpdateAll();
+    }
 }
